Some fixes for errors thrown by clang++ 18.1.8

+++ test/unit/example_sources/example_linear_reaction_CX.hpp
@@ -72,7 +72,7 @@ inline void linear_reaction_CX_example(ParticleGroupSharedPtr particle_group) {
           Sym<REAL>(prop_map[default_properties.tot_reaction_rate]), // The total reaction rate sym 
           Sym<REAL>(prop_map[default_properties.weight]), // The weight sym 
           projectile_species.get_id(), // Ingoing partice state id 
-          std::array<int, 1>{target_species.get_id()}, // State IDs of all the products - here just one
+          std::array<int, 1>{static_cast<int>(target_species.get_id())}, // State IDs of all the products - here just one
           rate_data, // Reaction data used for the rate calculation
           cx_kernel, // Reaction kernel object to be used
           particle_spec, // The reaction also needs to know the ParticleSpec of the group it is going to be applied to
